Ammo is unlimited, removed Gun reload
We could not fix the bug that upon a new scene being loaded, you would not be able to reload your gun, meaning the game was unplayable. So in order for this to be playable for the Game Jam I have removed ammo and the ability to reload. It should be easy to implement back in to the game when we return to this project.